% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdma2.R
\name{plot_summary_DMR}
\alias{plot_summary_DMR}
\title{Summary plot for wrap_mediation_DMR function}
\usage{
plot_summary_DMR(res_wrap_mediation_DMR, res_DMR_build)
}
\arguments{
\item{res_wrap_mediation_DMR}{result object from wrap_mediation_DMR() function}

\item{res_DMR_build}{result object from DMR_build() function}
}
\value{
Summary plot
}
\description{
This function draw a summary plot of the mediation analysis
}
\examples{

# see wrap_mediation_DMR example

}
\author{
Basile Jumentier
}
