% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdma2.R
\name{plot_summary_med}
\alias{plot_summary_med}
\title{Summary plot for wrap_mediation function}
\usage{
plot_summary_med(res_wrap_mediation)
}
\arguments{
\item{res_wrap_mediation}{result object from wrap_mediation() function}
}
\value{
Summary plot
}
\description{
This function draw a summary plot of the mediation analysis
}
\examples{

# see wrap_mediation example

}
\author{
Basile Jumentier
}
